fix:(HistoricoAgendamentoController/HistoricoAgendamentoService):kips agendamento
