+++ src/main/java/com/spring/ApiSystem/domain/agendamento/AgendamentoRepository.java
@@ -210,28 +210,4 @@ public interface AgendamentoRepository extends JpaRepository<Agendamento, Long> 
             @Param("status") AgendamentoStatus status
     );
 
-    @Query("""
-        SELECT
-            COALESCE(SUM(CASE WHEN a.status IN (
-                PENDENTE_PERSONAL_APROVACAO,
-                PENDENTE_PERSONAL_CONCLUIR
-            ) THEN 1 ELSE 0 END), 0) as totalPendente,
-    
-            COALESCE(SUM(CASE WHEN a.status IN (
-                PENDENTE_CLIENTE_APROVACAO,
-                APROVADO
-            ) THEN 1 ELSE 0 END), 0) as totalRespondido,
-    
-            COALESCE(SUM(CASE WHEN a.status IN (
-                CANCELADO_CLIENTE,
-                CANCELADO_PERSONAL
-            )
-            AND YEAR(a.data) = YEAR(CURRENT_DATE)
-            AND MONTH(a.data) = MONTH(CURRENT_DATE)
-            THEN 1 ELSE 0 END), 0) as totalCanceladoPorMesAtual
-        FROM agendamento a
-        WHERE a.personal.id = :personalId
-    """)
-    ResTotalAgendamentoByStatusProjection countTotalStatusAgendamentoByPersonal(@Param("personalId") Long personalId);
-
 }
