+++ src/main/java/com/spring/ApiSystem/shared/config/SecurityConfig.java
@@ -103,6 +103,7 @@ public class SecurityConfig {
                     ).hasAuthority("ROLE_PERSONAL");
 
                     auth.requestMatchers(
+                           "/api/historico-agendamento/total-status",
                             "/api/agendamentos/*/confirmar-conclusao",
                             "/api/agendamentos/ausencia",
                             "/api/agendamentos/consultoria-realizadas/*",
